# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Yap API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/groups": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "Get groups for a service body",
                "operationId": "3e63a010ca5ead4c216a420ebcff9fe5",
                "parameters": [
                    {
                        "name": "serviceBodyId",
                        "in": "query",
                        "description": "ID of the service body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "manage",
                        "in": "query",
                        "description": "Whether to include management data",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Groups retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "service_body_id": {
                                                "type": "integer"
                                            },
                                            "id": {
                                                "type": "integer"
                                            },
                                            "parent_id": {
                                                "type": "integer",
                                                "nullable": true
                                            },
                                            "data": {
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Groups"
                ],
                "summary": "Create a new group",
                "operationId": "315d542826296a29326e3688714a0f9e",
                "parameters": [
                    {
                        "name": "serviceBodyId",
                        "in": "query",
                        "description": "ID of the service body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "enabled": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Group created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "service_body_id": {
                                                "type": "integer"
                                            },
                                            "id": {
                                                "type": "integer"
                                            },
                                            "parent_id": {
                                                "type": "integer",
                                                "nullable": true
                                            },
                                            "data": {
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/groups/{id}": {
            "put": {
                "tags": [
                    "Groups"
                ],
                "summary": "Update a group",
                "operationId": "a259eaba26cab0da28f5b0a3095bf01c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the group to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "enabled": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Group updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "service_body_id": {
                                                "type": "integer"
                                            },
                                            "id": {
                                                "type": "integer"
                                            },
                                            "parent_id": {
                                                "type": "integer",
                                                "nullable": true
                                            },
                                            "data": {
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Groups"
                ],
                "summary": "Delete a group",
                "operationId": "13312c959b505531f2863efd956f944d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the group to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Group deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Group not found"
                    }
                }
            }
        },
...
        "/api/v1/volunteers/schedule": {
            "get": {
                "tags": [
                    "VolunteerSchedule"
                ],
                "summary": "Get volunteer helpline schedule",
                "operationId": "23b05f33df56a822428f50d42f88028a",
                "parameters": [
                    {
                        "name": "service_body_id",
                        "in": "query",
                        "description": "ID of the service body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Volunteer schedule retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "number": {
                                                "type": "string"
                                            },
                                            "gender": {
                                                "type": "string"
                                            },
                                            "responder": {
                                                "type": "boolean"
                                            },
                                            "type": {
                                                "type": "string"
                                            },
                                            "language": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            },
                                            "notes": {
                                                "type": "string"
                                            },
                                            "shift_info": {
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "BearerAuth": []
        }
    ]
}